+++ lib/generators/rails_template18f/sidekiq/sidekiq_generator.rb
@@ -17,26 +17,28 @@ module RailsTemplate18f
       end
 
       def configure_server_runner
-        append_to_file "Procfile.dev", "worker: bundle exec sidekiq"
+        append_to_file "Procfile.dev", "worker: bundle exec sidekiq\n"
         insert_into_file "manifest.yml", indent(<<~EOYAML), after: /processes:$\n/
           - type: worker
             instances: ((worker_instances))
             memory: ((worker_memory))
             command: bundle exec sidekiq
         EOYAML
+        insert_into_file "manifest.yml", "\n  - #{app_name}-redis-((env))", after: "services:"
         inside "config/deployment" do
           append_to_file "staging.yml", <<~EOYAML
             worker_instances: 1
-            web_memory: 256M
+            worker_memory: 256M
           EOYAML
           append_to_file "production.yml", <<~EOYAML
             worker_instances: 1
-            web_memory: 512M
+            worker_memory: 512M
           EOYAML
         end
       end
 
       def configure_active_job
+        generate "rails_template18f:cloud_gov_config"
         copy_file "config/initializers/redis.rb"
         application "config.active_job.queue_adapter = :sidekiq"
       end
@@ -54,7 +56,7 @@ module RailsTemplate18f
         boundary_filename = "doc/compliance/apps/application.boundary.md"
 
         insert_into_file boundary_filename, indent(<<~EOB, 16), after: /ContainerDb\(app_db.*$\n/
-          Container(worker, "<&layers> Sidekiq workers", "Ruby <%= @ruby_version %>, Sidekiq", "Perform background work and data processing")
+          Container(worker, "<&layers> Sidekiq workers", "Ruby #{ruby_version}, Sidekiq", "Perform background work and data processing")
           ContainerDb(redis, "Redis Database", "AWS ElastiCache (Redis)", "Background job queue")
         EOB
         insert_into_file boundary_filename, <<~EOB, before: "@enduml"